+++ src/main/java/no/nav/tag/tiltaksgjennomforing/avtale/EndreStillingsbeskrivelse.java
@@ -11,5 +11,5 @@ public class EndreStillingsbeskrivelse {
     Integer stillingStyrk08;
     Integer stillingKonseptId;
     Integer stillingprosent;
-    Integer antallDagerPerUke;
+    Double antallDagerPerUke;
 }
